+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/test/helper/GraphQLConfigurationTestHelper.java
@@ -45,6 +45,7 @@ public class GraphQLConfigurationTestHelper implements GraphQLConfiguration {
 	public boolean generateJPAAnnotation = true;
 	public boolean generateUtilityClasses = true;
 	public String javaTypeForIDType = GenerateServerCodeConfiguration.DEFAULT_JAVA_TYPE_FOR_ID_TYPE;
+	public String jsonGraphqlSchemaFilename = null;
 	public PluginMode mode = null;
 	public String packageName = "org.my.test.package";
 	private Integer maxTokens = Integer.parseInt(CommonConfiguration.DEFAULT_MAX_TOKENS);