+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-client/src/test/java/com/graphql_java_generator/samples/forum/test/jsonSchema/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains tests for the jsonSchemaFilename parameter.<br/>
+ * This parameter defines a json file that contains the result of an introspection query.
+ */
+package com.graphql_java_generator.samples.forum.test.jsonSchema;